Switch from Zustand to Jōtai
The dev seems to think Jōtai is a nicer library, and I wanted to play with it.

Refactor out GridItem so we only re-render when a specific album
changes, not when any in the list do.

Update tailwind and use new JIT features like shorthand opacity and caret colour

Support tab cycling in SearchBox